Change list state template to mut() by using class
Temporarily convert TTTListState to class from an actor in order to
support the template argument being a mut() rather than a pure function.

TTTListState really should be an actor but there seems to be some bug
where the compiler thinks we are leaking outside the actor seal, so
doing this temporarily. https://github.com/actonlang/acton/issues/1953