Support devices & RFS transforms
This is a first attempt at supporting devices and RFS transforms. There
is now an orchestron-rfs YANG model which has to be used as the RFS
layer in the system, i.e. the penultimate layer must contain the
orchestron-rfs model. This model has a list of device with the device
config, like addresses, credentials etc. There is also a /rfs list under
which it is expected that the user augments in their RFS services. This
list is keyed by the device name, which is the basic idea of how RFS
services will be run, they must be associated with one particular device
to effectively become part of a closed loop for that device. This is
important because we want to be able to rerun RFS services in a very
natural way if device parameters change, like the device
capabilities (after software upgrade etc).

In retrospect, after having written this, I think we should soften the
design a little bit, so instead of requiring the user to augment in
their services, they should be allowed to define their own model but it
must still conform to this basic principle, i.e. RFS services are nested
under a list of devices keyed by device name and then the system can
pick up from there.