Add SysSpec
We now have a SysSpec, which is the specification of the system, like
what laters and what device types are supported. From it, we generate
the code to be generated.

More importantly is the separation of the normal layers from device
types, so the bottom layer is now NOT device models anymore, or at least
not as specified in the list of layers. The main reason for this is that
for the normal layers, each layer is one YANG tree whereas for the
devices, we have separate trees, one for each device type.