Generate transform base class
For each transform we now generate a base class. The idea is that the
user, when implementing their actual transform, then inherits from the
base. This way, the type signature of the transform function can be
inherited, so we know the type without the user having typ explicitly
type it out (it's usually quite long). We can also add in other
attributes to the base transform class if we find that necessary in the future.

Further the output_root is known for each layer, so we now include it so
the users module can import even less.