Add device module capabilities
Now we fetch the capabilities from devices and parse them into a list of
modules that the device supports. This is accessible from the RFS
Transforms so we can apply condition on this module-set to determine
which device type we are dealing with.

It's not a good idea to parse NETCONF capabilities. We should support
the YANG library to fetch the list of supported modules. This is a quick
hack for now...