Add schema transform pipeline to device model builder
In some cases we may need to transform the device YANG modules schema before
building the adata classes. We expose a schema transform pipeline to users,
where they can specify a matching pattern and list schema transforms to be
applied to the modules. The results of the schema transforms are written to
disk so that changes may be tracked.