Don't write generated layers modules if no content changed
This is a poor-mans substitute for lack of content hashing. We don't write out
the generated files and change mtimes if the content is exactly the same.
Ideally we would look at mtimes / hashes of input (YANG) data and combine that
with the content hash of the code that generates adata and TTT modules, but
Acton doesn't have that yet ...