Distinguish success vs failure + handle more fails
Device disconnects now count as transaction_done but with a failure. I
don't think we run any risk of receiving multiple callbacks since we
only ever have one outstanding message to the NETCONF server. We can
however get _on_error() called separately, so need to make sure that
reconnects are not racy between sources of this information (but we
don't reconnect now, so no problems :P)