Modular device types
This is the first step towards modular device types. The Device-actor is
now an abstract device and all the protocol specific stuff is moved into
a DeviceAdapter class, or rather it is expected that subclasses are used
to implement specific Device Adapters. NetconfAdapter is the only real
one we currently have, which is just the same NETCONF code as before but
moved over.

Since we don't have actor inheritance, we end up with a NetconfDriver
actor doing the actual work and DeviceAdapter being an in-between class
definition. Once we have actor inheritance, I imagine NetconfDriver
should be renamed to NetconfAdapter and replace the current class.

I did consider having the Device-actor as an abstract type (ignore the
fact we don't have actor inheritance yet) and then NetconfDevice would
be a concrete implementation.. but ultimately decided against that and
in benefit of this two-actor approach. It is possible to change the
device type and thus which adapter is used. We also want to maintain
state over time, like the intended configuration. These goals are
incompatible with using a single actor, since replacing it would mean
loosing state. Thus I think it's cleaner to have the abstract
Device-actor that always exists and can maintain state over time, which
then delegates specific tasks to the adapter implementation. The adapter
then becomes pretty much stateless, which simplifies it whereas
complicated state-keeping code in the Device-actor stays abstract and
thus reusable for all devices. Pretty nice!