Use modset_id for config integrity
It is crucial that the configuration produced by RFS transforms are
based on the correct modset of the device. Initially before we have
connected to the device, the modset is empty, yet the RFS transform must
run. We allow this but pass None for the config and None for modset_id,
which means that we only wish to register interest in this device. The
expectation is that once we have the device proper modset, a reconf is
triggered from which the RFS transforms now can compute a correct
configuration.

All configuration passed down to the DeviceMgr is checked for modset_id
compliance so that we don't get any time skew related problems, like if
.configure() calls are in-transit while the modset is changing or
similar.

If we are running in mock mode, i.e. no WorldCap is provided, the
default adapter is now MockAdapter (which is actually the behavior
previous to the introduction of device adapters). This is probably
always a good choice, even later on when we make the choice of adapter
more explicit, since it means we'll always support mocking out of the
box.