Restore list element Container key attribute
After a list element gdata.Container is read from TTT with .get(), it is
missing the key attribute. This makes the gdata structure invalid, and
results in errors in places where this data is used.

This is a temporary fix until we move away from the key attribute
altogether and replace it with a method to construct the key on demand
by reading the child key leaf values.