Attempt to constrain TTT Node composition with types
Handle the case where a ttt._ContainerTransaction that tracks changes to
embedded transforms in the list element, is also made aware of the key
leaves for its list. These are *not* part of any embedded transforms,
but are still the responsibility of the list element container. This
means that ttt._ContainerTransaction must be provided with key_names
from its containing list.

I tried expressing this with types and was somewhat successful - this
case is now handled by the stack
`ttt.ListStructural(ttt.ListElement(...), <key_names>)`. But what is
missing is a constraint that would prevent
`ttt.List(ttt.Container(...), <key_names>)` and only allow to ttt.List
to contain "user-facing" transforms, a result of "orchestron:X"
extension statements.