Run <get-config> after connect / commit
We grab a copy of the running config from the device after a successful
connection and commit. At the moment this config is stored as XML and
not as parsed gdata. This will change soon when we start using the new
DeviceType class to push parser function references to the Device actor though.

In the Device actor we also compare the result of get-config (new) with
the currently known (old) running config and log the diff.