Incorporate device-type
We have taken som preparatory steps towards having device-types. This
pretty much completes it, at least an initial version. Each device-type
now has an adapter-type which the build gen populates in sysspec with
the currently only supported real adapter - the NetconfAdapter. There
are no actual working adapters other than NetconfAdapter so we can't
truly test this out, but it should work. We'll have to extend the build
gen once we actually have others.

The list of device-types is injected to the DeviceManager from sysspec
and then we instantiate whichever is configured. We still default to the
NetconfAdapter in case no other device-type is configured, but I think
this is really just temporary, we probably want to pick a device-type so
we know what models are supported by the device, or at least our
expectation of supported models. We can still add in automatic selection
of device-type / schemas based on support modules of the device and some
heuristics.