Use NoAdapter per default & add init_dmc
init_dmc is the devicemetaconfig that we feed in when the adapter is
constructed - having access to early is nicer, although I suppose we
mostly won't be able to do anything with it since trying to initialize
any client early on is likely to result in cyclic problems between
functions and actor attributes.

We now use the NoAdapter per default, which really does nothing. As soon
as DMC is set, we'll create the proper adapter!