Rename functions for getting config
I think this is easier to understand. The DeviceAdapter is supposed to
keep a copy of the device config available. .get_config will get that
cached version of the config or we have fetch_config which will actually
go and retrieve the config from the device. It's just a little
misaligned with netconf that calls it get-config.