Proper NETCONF commit procedure
This implements a better commit procedure in the NetconfAdapter. It
currently only changes the procedure itself, going from a simple
edit-config + potential commit, to now being a rather elaborate dance
of multiple steps. Crucially, one of the steps is to fetch the
current running configuration from the device, which enables us to
compute a diff between the current devie config and our target config.
Diffs are more efficient, but more importantly, they allow us to remove
things configured on the device, which are not in our target config,
i.e. that should not be there. However, seeing how we are not quite
ready to remove stuff, partly due to weaknesses in our own diff etc
functions and partly in the downstream code (sorespo), the diff is
currently disabled. The rest of the more robust procedure is in place
though and is a good improvement.

The error handling in DeviceMgr is also improved with more granular
errors under the two top categories Transient / Permanent Error. We now
schedule retries based on this information which should make us recover
from more situations, like when a device is currently locked (from
someone elses session or during its startup).

See the description on the NetconfDriver for more details on its
behavior.