Skip device test
Bah, this doesn't feel very good, but we just don't have very good means
of controlling entire test environments including external things, so
it'll have to do for now.

To actually run the test, modify the global run_test = True and start a
device we can test with on the expected port etc.