Add DeviceMgr.reconfigure()
DeviceMgr.reconfigure() reconfigures a device. This can be called in
order to ensure that the device is configured in accordance with the
current target configuration. While configuration changes from within
Orchestron naturally trigger configuration of the device, it is possible
that the device drifts out of sync, for example due to manual changes on
the device. Calling reconfigure() will push the current target
configuration to the device again.