Rework commit procedure
The DeviceMgr is now equipped with a view of the running configuration
of the device. Rather than blindly sending configuration from the
DeviceMgr to the DeviceAdapter, the DeviceMgr now computes a diff and
sends the diff to the DeviceAdapter, in addition to the full
configuration as well as its view of the currently running configuration
on the device. This was actually part of my very initial design for
DeviceMgr (although never really implemented, just as a placeholder),
but I removed it to keep things simpler when implementing the
DeviceAdapter commit procedure. We're now bringing running_conf back to
the DeviceMgr with the primary reason being so that we can produce
meaningful diffs in the DeviceMgr. This in turn is driven by the need to
inspect, review and approve individual changes. I attempted to implement
such approval queues and made this realization, which has triggered this
work. Once this is in place, we are now in a place where we can build
approval queues on top. We are sending both a diff as well as the full
targe config and our current view of the device's running config, so
that the DeviceAdapter can use the precomputed diff but also have the
ability to compute a diff of its own, in cases where that is necessary,
essentially for devices not following NETCONF / YANG semantics in how
diffs are applied etc. Or for applying magic haxx, like our config_fixer
which automagically copies over certain configuration for IOS XR to
remain working in our lab environments.

Many things have improved in this branch. We have txid in the updated
version of gdata. We use this to attach a txid when we fetch
configuration from the device. For JUNOS we use the commit-seconds
metadata information as txid, which can be fetched rather quickly, while
for other NETCONF devices we fall back to computing a hash of the entire
configuration and using that as txid. Wildly inefficient, but with the
basic concept in place, I think we are in a good place to improve with
more efficient methods for specific devices depending on what they
support.

Since we now have a closed loop for device config reconciliation we also
have the opportunity for infinite loops! But we take extra precautions
to break such loops. The basic scenario for infinite loops is that we
compute a diff and send to the device. Then something happens which
triggers us (well we don't really have such automatic triggers yet) to
do a resync, which triggers us to recheck if the device running config
is the same as our target config, if not, send a new diff. So if the
device mutates the configuration or there are other reasons for why the
retrieved configuration is different from our target, we will get a loop!
There are currently three such cases that occur and basically forms the
basis for a loop:
- JUNOS adds in a `<version>` leaf in the configuration
- we write passwords and secrets as plain text but get them back in
  hashed or encrypted form, which the diff is unaware of and thus wants
  to correct them back to their expected plain-text version
- parsing issues where unions of strings & int mean we parse values as
  string which should really be integers and in the diff, a comparison
  of a str and int will result in a diff

We detect loops by saving the last successfully sent diff and comparing
any newly computed diff to the last_sent_diff. If they are the same, it
means we have already attempted to configure the device with this exact
output and doing so again (and expecting a different outcome) would be
foolish and so we stop.

The DeviceAdapter.configure() method is now expected to return (via
callback) the resulting configuration on the device. There are
essentially two ways by which to get the resulting configuration; either
we say that it is deterministic, so we can get the configuration on the
device by computing it in Orchestron and for NETCONF / YANG this is
conceptually true, like if we apply a diff on our local view of the
device's running config and do it on the device itself, we should end up
with the same result... were it not for those cases where the device
itself mutates the config. So in practice, the pessimistic approach is
instead to re-fetch the configuration from the device after editing the
config. We currently stick to the optimistic view, but doing a resync
will sort of expose the true nature of things.