Be pessimistic about pending_txids
Previously we've taken the content of pending_txids and put in
current_txids when we start a configuration round, leaving pending_txids
empty. If the configuration attempt, we put back all the items from
current_txids to pending_txids.

Now, we instead don't empty pending_txids when we start a configuration
round, rather we copy it over into current_txids and then, only on a
successful configuration round, we remove the txids in current_txids
from pending_txids. We arrive at the same result, but sort of take a bit
more pessimistic view and "leaning" more towards that configuration
attempts can fail. I don't actually think that's very likely, but with
the upcoming changes for config queue etc, I think this better aligns
with those semantics and will make the code easier to read.