Track reconf_txids separately
Rather than track reconf intents as any other transaction in
pending_txids, we now track them separately. This means we can be a
little bit more precise, if a configuration attempt fails, we used to
inspect pending_txids to see if it had changed, which indicates there is
new configuration. Normally this is true, for new configuration means we
write the latest received configuration to target_conf and add the new
tid to pending_txid. However, for a reconf, we would only update
pending_txids, not touching target_conf. When we inspect pending_txids
to find out if there is new configuration, those reconf intents will
trick us. Now that we track reconf_txids separately, an updated
pending_txids will accurately capture if the configuration has actually
changed!