Add confq
New configuration sent from the transaction engine to the DeviceMgr via
its .configure() method is now added to a queue of outgoing
configuration. We consume from the queue when sending to the device. The
same aggregating batch semantics as before are used, that is, when
there are multiple ConfigItems in the confq, we send all to the device
at once, not one by one.

pending_txids is effectively replaced by the confq since it already
tracks pending tids and the other use of pending_txids was already
refactored to use reconf_txids.