Simplify tids terminology
We have used tid(s) and txids interchangeably. We recently added the
concept of device txid, which increases the chance of confusion. To
simplify, we use tid and tids for our internal transaction ids. Device
txid is based on concepts from a netmod draft which calls it txid, so
that's why it remains as device_txid (and we try to be explicit calling
it device txid where we can).