Use resulting config from adapter
Unconditionally use the resulting configuration from the adapter. This
was a remnant from my first idea around resulting config from the
adapter where I thought it was an optional part of the adapter API but
have since changed my mind - the adapter MUST return the resulting
configuration of the device and so the DeviceMgr should always use it.