Add config q item to reconcile after resync
In a resync after we discover that the device configuration is different
than our local view of it, we want to reconcile (well, overwrite) the
device. When approval_required is set we can't just do that, so we push
a new configuration item to the head of the config queue to update the
config to what we want it to be (target_conf) so the user is given the
chance to explicitly approve it.