Add SchemaRegistry
Fetch device schemas on connect, compile and cache the result based on a
content hash of the YANG source files. By using a single SchemaRegistry
actor the DeviceMgr actors may synchronize on compilation and share a
single compiled schema tree for devices with identical YANG.

If the DeviceMgr actor reconnects to the device in the middle of an
ongoing schema download, it will create a new SchemaGetter instance. But
the old one is still running and will execute callbacks. To avoid
processing stale responses, we store a reference to the active
SchemaGetter actor and then compare it with the provided reference in
the callback.