Add minisys
This is a system built on the Orchestron platform, just like SORESPO and
other systems we've built. This one is smaller, like really small, thus
"mini". The benefit is of course to have it right here in the Orchestron
repo so we can run tests on it within the same git repo, simplifying
like everything.

We have extremely simply transforms and just a single device type, the
"IETF device", which basically just supports the ietf-system.yang
model (and its dependencies).