Add script for checking build.act.json consistency
This repository includes 4 Acton projects, each with their own package
dependencies declared in build.act.json. We do have the "pkg-upgrade"
recipe that updates the package dependencies to the latest version
available for all projects, but it is very easy to miss something if
doing a manual targeted update. We execute the "check-dep-consistency"
recipe in CI to ensure all versions are in sync.