Preserve list element key leaves in ListState
For the case when a list element is not an embedded transform but just a
_Container (and one of its child nodes is a list with an embedded
transform), ListState must keep track of the key leaf values.
Previously this was a bit of a hack with reconstructing the gdata.Leaf
nodes by splitting the concatenated list element key (string).

We keep these values in the ListState actor and not just in
_ListTransaction to also make them available in _List.get(). That
existence of transaction bypass functions _List.get() (and
_Container.get()) is questionable and they may well be removed in the
future. At that point we can refactor and move key leaf restoring to _ListTransaction.