Split transpose into transpose_list, transpose_container
Initially transpose was a single function that always returned
`dict[str, dict[str, gdata.Node]]`. The inner dict str key values represented:

1. container child node name when transposing a container,
2. list element key (concatenated keys, comma separated) when transposing a list,
3. WILDKEY empty string when transposing Absent.

After the gdata.Id migration the container children (1.) are of type
gdata.Id. Since value, the common type to gdata.Id and str, is not
Hashable, it cannot be used as a dict key. There are multiple solutions
to this though:

1. With union type support in Action, we could express this in a single
function, like `dict[str|gdata.Id, ...]`.
2. By storing the list element key string in gdata.Id (just name, no
qualifier), we could use the gdata.Id key for both cases.
3. Create separate functions for transposing containers, lists.

I prefer 3 over 2 because the semantics of the outer dict key after
transposing are different for containers vs. lists: for containers it is
used as an lookup key as well as restored when reading data from TTT,
for list elements it is only a lookup key (in the past was also used to
restore key leaves, but that was incorrect).